Fix a signedness issue
QML does not support QList<unsigned int>. Declaring
a public method taking such an object as a parameter
lead to receiving an empty QList when called from QML
(even though unit tests ran fine).
The easiest solution was to drop unsigned int in favor
of (signed) int; positive value checks are implemented anyway.

+++ tests/scheduling/tst_schedulingtest.cpp
@@ -11,18 +11,18 @@ SchedulingTest::SchedulingTest()
 
 void SchedulingTest::testCannotCreate_data()
 {
-    QTest::addColumn<unsigned int>("machineDuration");
-    QTest::addColumn<QList<unsigned int> >("jobDurations");
+    QTest::addColumn<int>("machineDuration");
+    QTest::addColumn<QList<int> >("jobDurations");
 
-    QTest::newRow("no_machines") << (uint)0 << QList<unsigned int>({1, 2, 3, 4, 5, 6, 7, 8, 9, 10});
-    QTest::newRow("no_jobs") << (uint)3 << QList<unsigned int>();
-    QTest::newRow("zero_job") << (uint)10 << QList<unsigned int>({0});
+    QTest::newRow("no_machines")    << 0  << QList<int>({1, 2, 3, 4, 5, 6, 7, 8, 9, 10});
+    QTest::newRow("no_jobs")        << 3  << QList<int>();
+    QTest::newRow("zero_job")       << 10 << QList<int>({0});
 }
 
 void SchedulingTest::testCannotCreate()
 {
-    QFETCH(unsigned int, machineDuration);
-    QFETCH(QList<unsigned int>, jobDurations);
+    QFETCH(int, machineDuration);
+    QFETCH(QList<int>, jobDurations);
 
     TaskScheduler scheduler;
 
@@ -45,19 +45,19 @@ void SchedulingTest::testCannotCreate()
 
 void SchedulingTest::testCanSchedule_data()
 {
-    QTest::addColumn<unsigned int>("machineDuration");
-    QTest::addColumn<QList<unsigned int> >("jobDurations");
+    QTest::addColumn<int>("machineDuration");
+    QTest::addColumn<QList<int> >("jobDurations");
 
-    QTest::newRow("more_machines") << (uint)15 << QList<unsigned int>({1, 2, 3, 4, 5, 6, 7, 8, 9, 10});
-    QTest::newRow("more_jobs") << (uint)3 << QList<unsigned int>({1, 2, 3, 4, 5, 6, 7, 8, 9, 10});
-    QTest::newRow("single_machine") << (uint)1 << QList<unsigned int>({1, 2, 3, 4, 5, 6, 7, 8, 9, 10});
-    QTest::newRow("large_job") << (uint)10 << QList<unsigned int>({150});
+    QTest::newRow("more_machines")  << 15 << QList<int>({1, 2, 3, 4, 5, 6, 7, 8, 9, 10});
+    QTest::newRow("more_jobs")      << 3  << QList<int>({1, 2, 3, 4, 5, 6, 7, 8, 9, 10});
+    QTest::newRow("single_machine") << 1  << QList<int>({1, 2, 3, 4, 5, 6, 7, 8, 9, 10});
+    QTest::newRow("large_job")      << 10 << QList<int>({150});
 }
 
 void SchedulingTest::testCanSchedule()
 {
-    QFETCH(unsigned int, machineDuration);
-    QFETCH(QList<unsigned int>, jobDurations);
+    QFETCH(int, machineDuration);
+    QFETCH(QList<int>, jobDurations);
 
     TaskScheduler scheduler;
 
@@ -77,19 +77,19 @@ void SchedulingTest::testCanSchedule()
 
 void SchedulingTest::testCanScheduleMultipleTimes_data()
 {
-    QTest::addColumn<unsigned int>("machineDuration");
-    QTest::addColumn<QList<unsigned int> >("jobDurations");
+    QTest::addColumn<int>("machineDuration");
+    QTest::addColumn<QList<int> >("jobDurations");
 
-    QTest::newRow("more_machines") << (uint)15 << QList<unsigned int>({1, 2, 3, 4, 5, 6, 7, 8, 9, 10});
-    QTest::newRow("more_jobs") << (uint)3 << QList<unsigned int>({1, 2, 3, 4, 5, 6, 7, 8, 9, 10});
-    QTest::newRow("single_machine") << (uint)1 << QList<unsigned int>({1, 2, 3, 4, 5, 6, 7, 8, 9, 10});
-    QTest::newRow("large_job") << (uint)10 << QList<unsigned int>({150});
+    QTest::newRow("more_machines")  << (uint)15 << QList<int>({1, 2, 3, 4, 5, 6, 7, 8, 9, 10});
+    QTest::newRow("more_jobs")      << (uint)3  << QList<int>({1, 2, 3, 4, 5, 6, 7, 8, 9, 10});
+    QTest::newRow("single_machine") << (uint)1  << QList<int>({1, 2, 3, 4, 5, 6, 7, 8, 9, 10});
+    QTest::newRow("large_job")      << (uint)10 << QList<int>({150});
 }
 
 void SchedulingTest::testCanScheduleMultipleTimes()
 {
-    QFETCH(unsigned int, machineDuration);
-    QFETCH(QList<unsigned int>, jobDurations);
+    QFETCH(int, machineDuration);
+    QFETCH(QList<int>, jobDurations);
 
     TaskScheduler scheduler;
